requirements.txt
flask==3.0.2
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.5 